AndroidAppUploader/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
daphne==2.2.5
Django==2.1.5
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
psycopg2==2.7.7
PyHamcrest==1.9.0
pytz==2018.9
PyJWT==1.7.1
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
websocket-client==0.54.0
zope.interface==4.6.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.1
requests-oauthlib==1.2.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
